Policy Details Store refactor into multiple modules (#112697)
