+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/action/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { PolicySettingsAction } from './policy_settings_action';
+import { PolicyTrustedAppsAction } from './policy_trusted_apps_action';
+
+export type PolicyDetailsAction = PolicySettingsAction | PolicyTrustedAppsAction;
