+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/index.ts
@@ -11,8 +11,7 @@ import { AppAction } from '../../../../../common/store/actions';
 import { Immutable } from '../../../../../../common/endpoint/types';
 
 export { policyDetailsMiddlewareFactory } from './middleware';
-export { PolicyDetailsAction } from './action';
-export { policyDetailsReducer } from './reducer';
+export { policyDetailsReducer, initialPolicyDetailsState } from './reducer';
 
 export interface EndpointPolicyDetailsStatePluginState {
   policyDetails: Immutable<PolicyDetailsState>;
@@ -21,3 +20,4 @@ export interface EndpointPolicyDetailsStatePluginState {
 export interface EndpointPolicyDetailsStatePluginReducer {
   policyDetails: ImmutableReducer<PolicyDetailsState, AppAction>;
 }
+export { PolicyDetailsAction } from './action';
