+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/middleware/policy_trusted_apps_middleware.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { MiddlewareRunner } from '../../../types';
+
+export const policyTrustedAppsMiddlewareRunner: MiddlewareRunner = async (
+  coreStart,
+  store,
+  action
+) => {
+  // FIXME: implement middlware for trusted apps
+};
