+++ x-pack/plugins/security_solution/public/management/pages/policy/store/policy_details/selectors/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export * from './policy_settings_selectors';
+export * from './trusted_apps_selectors';
